requirements.txt
albucore==0.0.13
albumentations==1.4.10
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
astor==0.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.8
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
exceptiongroup==1.2.2
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
fonttools==4.55.3
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.19.1
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.3
pillow==11.0.0
protobuf==3.20.2
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyclipper==1.3.0.post6
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-docx==1.1.2
pyttsx3==2.98
pytz==2024.2
pywin32==308
PyYAML==6.0.2
RapidFuzz==3.11.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.12.12
toml==0.10.2
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.55
ultralytics-thop==2.0.13
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
